+++ package.json
@@ -5,11 +5,10 @@
     "type": "git",
     "url": ""
   },
-  "type": "module",
   "description": "REST service starter for RS School",
   "main": "src/server.ts",
   "scripts": {
-    "start": "cross-env nodemon --watch './src/**/*.ts' ./src/server.ts",
+    "start": "cross-env nodemon --watch './src/**/*.ts' --exec ts-node ./src/server.ts",
     "lint": "eslint --ext .js,.ts --fix src/",
     "test": "cross-env DEBUG=rs:* jest --testMatch \"<rootDir>/test/e2e/test/*.test.js\" --noStackTrace --runInBand",
     "test:auth": "cross-env DEBUG=rs:* TEST_MODE=auth jest --noStackTrace",
@@ -57,6 +56,8 @@
     "@types/swagger-ui-express": "^4.1.2",
     "@types/uuid": "^8.3.0",
     "@types/yamljs": "^0.2.31",
+    "@typescript-eslint/eslint-plugin": "^4.25.0",
+    "@typescript-eslint/parser": "^4.25.0",
     "chai": "4.3.4",
     "debug": "4.3.1",
     "dirty-chai": "2.0.1",
@@ -73,6 +74,7 @@
     "nodemon": "2.0.7",
     "prettier": "2.2.1",
     "supertest": "6.1.3",
-    "ts-node": "^10.0.0"
+    "ts-node": "^10.0.0",
+    "typescript": "^4.3.2"
   }
 }
