+++ src/common/config.ts
@@ -1,4 +1,4 @@
-const dotenv = require('dotenv');
+import dotenv from 'dotenv';
 import path from 'path';
 
 dotenv.config({
@@ -6,9 +6,9 @@ dotenv.config({
 });
 
 export default {
-  PORT: process.env['PORT'],
-  NODE_ENV: process.env['NODE_ENV'],
-  MONGO_CONNECTION_STRING: process.env['MONGO_CONNECTION_STRING'],
-  JWT_SECRET_KEY: process.env['JWT_SECRET_KEY'],
-  AUTH_MODE: process.env['AUTH_MODE'] === 'true'
+  PORT: process.env.PORT,
+  NODE_ENV: process.env.NODE_ENV,
+  MONGO_CONNECTION_STRING: process.env.MONGO_CONNECTION_STRING,
+  JWT_SECRET_KEY: process.env.JWT_SECRET_KEY,
+  AUTH_MODE: process.env.AUTH_MODE === 'true'  
 };
