+++ src/resources/boards/board.memory.repository.ts
@@ -1,4 +1,5 @@
 import Board from "./board.model";
+
 const boards: Board[] = [];
 
 /**
