+++ src/resources/boards/board.model.ts
@@ -2,8 +2,11 @@ import { v4 } from 'uuid';
 
 class Board {
   id: string;
+
   title: string;
+
   columns: never[];
+
   constructor({
     id = v4(),
     title = 'BOARD',
