+++ src/resources/boards/column.model.ts
@@ -2,8 +2,11 @@ import { v4 } from 'uuid';
 
 class Column {
   id: string;
+
   title: string;
+
   order: number;
+
   constructor({
     id = v4(),
     title = 'COLUMN',
