+++ src/resources/tasks/task.model.ts
@@ -2,12 +2,19 @@ import { v4 } from 'uuid';
 
 class Task {
   id: string;
+
   title: string;
+
   order: number;
+
   description: string;
+
   userId: string | null;
-  boardId: string;
+
+  boardId: string | undefined;
+
   columnId: string;
+
   constructor({
     id = v4(),
     title = 'TASK',
