+++ src/resources/users/user.router.ts
@@ -1,6 +1,7 @@
 import express, { Request, Response } from 'express';
 import User from './user.model';
 import usersService from './user.service';
+
 const router = express.Router();
 
 router.route('/').get(async (res: Response) => {
@@ -9,7 +10,7 @@ router.route('/').get(async (res: Response) => {
 });
 
 router.route('/:id').get(async (req: Request, res: Response) => {
-  const user: User | undefined = await usersService.getById(req.params['id']);
+  const user: User | undefined = await usersService.getById(req.params.id);
   res.json(User.toResponse(user));
 });
 
@@ -22,13 +23,13 @@ router.route('/').post(async (req: Request, res: Response) => {
 
 router.route('/:id').put(async (req: Request, res: Response) => {
   const { body } = req;
-  const user = new User({ id: req.params['id'], ...body});
+  const user = new User({ id: req.params.id, ...body});
   const newUser = await usersService.update(user);
   res.json(newUser);
 });
 
 router.route('/:id').delete(async (req: Request, res: Response) => {
-  await usersService.remove(req.params['id']);
+  await usersService.remove(req.params.id);
   res
     .status(200)
     .json(null);
