+++ src/server.ts
@@ -1,5 +1,7 @@
-import PORT from './common/config';
+import OPTIONS from './common/config';
 import app from './app';
+
+const { PORT } = OPTIONS;
 
 app.listen(PORT, () =>
   console.log(`App is running on http://localhost:${PORT}`)
